project/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
bleach==6.2.0
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.2.1
cmudict==1.0.33
coloredlogs==15.0.1
Cython==3.1.2
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
Distance==0.1.3
distro==1.9.0
docopt==0.6.2
durationpy==0.10
emoji==1.2.0
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-vision==3.10.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
hangul-jamo==1.0.1
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipython==8.12.3
jamo==0.4.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kollocate==0.0.2
koparadigm==0.10.0
kss @ git+https://github.com/hyunwoongko/kss.git@ffd3176c64aef7fef8e781edb55578cceee4ee3a
kubernetes==33.1.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-huggingface==0.3.1
langchain-openai==0.3.28
langchain-pinecone==0.2.11
langchain-tests==0.3.20
langchain-text-splitters==0.3.9
langsmith==0.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mmh3==5.1.0
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.5
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.93.3
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.11.1
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pecab==1.0.8
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pinecone==7.3.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
pipreqs==0.5.0
platformdirs==4.3.8
pluggy==1.6.0
posthog==5.4.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pybind11==2.9.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.4.1
pytest-asyncio==0.26.0
pytest-benchmark==5.1.0
pytest-codspeed==4.0.0
pytest-recording==0.13.4
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-mecab-ko==1.3.7
python-mecab-ko-dic==2.1.1.post2
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.47.2
sympy==1.14.0
syrupy==4.9.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5.1
tossi==0.3.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.0
triton==3.3.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vcrpy==7.0.0
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
Whoosh==2.7.4
wrapt==1.17.2
xlrd==1.2.0
yarg==0.1.9
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
